splitted and improved message center meta
