+++ meta/messageCenter/index.ts
@@ -1,0 +1,4 @@
+export type { Message } from './message'
+export type { MessageTopic } from './messageTopic'
+
+export { MessageTopicStatus } from './messageTopic'
